\name{colorbar.sig}   
\alias{colorbar.sig} 
\title{Generates a  colour bar for spatial significance plots}
\description{This function generates a colour bar for the visualisation of significance of spatial bias.} 
\usage{colorbar.sig(color.lim=c(-3,3))}
\arguments{\item{color.lim}{limits for  color bar}}
\details{The function \code{colorbar.sig} produces a colour bar for 2D-plots generated by \code{sigxy.plot}.
         The colours used range from green (for the lower limit of the colour range)
         to red (for its upper limit).
         For visualisation, values below or above the limits for the colour range 
         (as given by \code{color.lim}) are set
         to the lower or upper limit, respectively. The function \code{colorbar.sig}
         is similar to  more general function \code{colorbar.mxy}. 
         It differs, however, in its axis annotation. Since it is used to present the significance
         in a log10-scale, annotation of axis tacks consists of negative values in both direction.}
\author{Matthias E. Futschik (\url{http://itb.biologie.hu-berlin.de/~futschik})}
\seealso{\code{\link{sigxy.plot}}, \code{\link{colorbar.mxy}}}
\keyword{hplot}



