\name{sw.olin}
\docTye{data}
\alias{sw.olin}
\title{Normalised cDNA microarray data of SW480/SW620 experiment}
\description{The data set \code{sw.olin} is derived from data set \code{sw}
	     by optimised local intensity-dependent normalisation (OLIN).}
\usage{data(sw.olin)}
\format{An object of class \dQuote{marrayNorm}}
\source{The original data (\code{sw}) was produced  and  provided by S.Pattison of the Cancer Genetics lab and A. Jeffs of
the Otago Genomics facility of the University of Otago, Dunedin, New Zealand.} 
\references{M. Futschik, A.Jeffs, S.Pattison, N.Kasabov, M.Sullivan, A.Merrie and A.Reeve, Gene expression profiling of metastatic and non-metastatic colorectal cancer cell-lines, \emph{Genome Letters}, vol.1, No.1, pp. 26-34, 2002}
\seealso{\code{\link{sw}}, \code{\link{olin}}}
\keyword{datasets}



