\name{sw.xy}
\docTye{data}
\alias{sw.xy}
\title{Spatial coordinates of spot locations of  SW480/SW620 experiment }
\description{The data set \code{sw.xy} contains the x- and y-coordinates of the spots
             in the  data set \code{sw}}
\usage{data(sw.xy)}
\format{A list of two matrices}
\source{The original data (\code{sw}) was produced  and  provided by S.Pattison of the Cancer Genetics lab and A.Jeffs of
the Otago Genomics facility of the University of Otago, Dunedin, New Zealand.} 
\references{M. Futschik, A.Jeffs, S.Pattison, N.Kasabov, M.Sullivan, A.Merrie and A.Reeve, Gene expression profiling of metastatic and non-metastatic colorectal cancer cell-lines, \emph{Genome Letters}, vol.1, No.1, pp. 26-34, 2002}
\seealso{\code{\link{sw}}}
\keyword{datasets}



